refactor: 폴더명 변경 및 주소 변경
