+++ src/components/ReviewComponent/AINoteChips.tsx
