+++ src/components/ReviewComponent/TotalScoreComponent.tsx
